Refactor psi declarations into dedicated header
