+++ XFoil.h
@@ -50,7 +50,7 @@ Harold Youngren. See http://raphael.mit.edu/xfoil for more information.
 #include "model/math_util.hpp"
 #include "model/spline.hpp"
 #include "model/boundary_layer.hpp"
-#include "model/psi.hpp"
+#include "model/psi_result.hpp"
 #include "xfoil_params.h"
 
 using namespace std;
@@ -251,12 +251,9 @@ class XFoil {
   bool mrchue();
   double calcHtarg(int ibl, int is, bool wake);
   Matrix2Xd ncalc(Matrix2Xd point, VectorXd spline_length, int n);
-
-  // TODO modle/psi.hppへ移植
-  PsiResult psilin(int i, Vector2d point, Vector2d normal_vector, bool siglin);
-  PsiResult psisig(int iNode, int jNode, Vector2d point, Vector2d normal_vector);
-  PsiResult psi_te(int i, Vector2d point, Vector2d normal_vector);
   
+#include "model/psi.hpp"
+
   PsiResult pswlin(int i, Vector2d point, Vector2d normal_vector);
   bool qdcalc();
   bool qiset();
